# Embeddings: eliminate bounds checks and unroll