--- a/enterprise/internal/embeddings/similarity_search.go
+++ b/enterprise/internal/embeddings/similarity_search.go
@@ -211,7 +211,16 @@ func CosineSimilarity(row []int8, query []int8) int32 {
 		panic("mismatched vector lengths")
 	}
 
-	for i := 0; i < count; i += 1 {
+	i := 0
+	for ; i+3 < count; i += 4 {
+		m0 := int32(row[i]) * int32(query[i])
+		m1 := int32(row[i+1]) * int32(query[i+1])
+		m2 := int32(row[i+2]) * int32(query[i+2])
+		m3 := int32(row[i+3]) * int32(query[i+3])
+		similarity += (m0 + m1 + m2 + m3)
+	}
+
+	for ; i < count; i++ {
 		similarity += int32(row[i]) * int32(query[i])
 	}