This checks the lengths of the input slices ahead of time so the compiler doesn't need to insert panics on slice accesses.

Bounds check elimination: 
```
name                               old time/op    new time/op    delta
SimilaritySearch/numWorkers=1-10      559ms ± 0%     499ms ± 0%  -10.74%  (p=0.016 n=4+5)
SimilaritySearch/numWorkers=2-10      283ms ± 0%     253ms ± 0%  -10.65%  (p=0.008 n=5+5)
SimilaritySearch/numWorkers=4-10      146ms ± 0%     130ms ± 0%  -10.52%  (p=0.008 n=5+5)
SimilaritySearch/numWorkers=8-10     73.7ms ± 0%    66.0ms ± 0%  -10.45%  (p=0.008 n=5+5)
SimilaritySearch/numWorkers=16-10    83.6ms ± 2%    75.0ms ± 3%  -10.29%  (p=0.008 n=5+5)
```

Loop unrolling:
```
name                               old time/op    new time/op    delta
SimilaritySearch/numWorkers=1-10      499ms ± 0%     439ms ± 0%  -12.12%  (p=0.008 n=5+5)
SimilaritySearch/numWorkers=2-10      253ms ± 0%     222ms ± 0%  -12.21%  (p=0.008 n=5+5)
SimilaritySearch/numWorkers=4-10      130ms ± 0%     115ms ± 0%  -11.99%  (p=0.008 n=5+5)
SimilaritySearch/numWorkers=8-10     66.0ms ± 0%    57.9ms ± 0%  -12.20%  (p=0.008 n=5+5)
SimilaritySearch/numWorkers=16-10    75.0ms ± 3%    65.7ms ± 2%  -12.36%  (p=0.008 n=5+5)
```

Both:
```
name                               old time/op    new time/op    delta
SimilaritySearch/numWorkers=1-10      559ms ± 0%     439ms ± 0%  -21.55%  (p=0.016 n=4+5)
SimilaritySearch/numWorkers=2-10      283ms ± 0%     222ms ± 0%  -21.56%  (p=0.008 n=5+5)
SimilaritySearch/numWorkers=4-10      146ms ± 0%     115ms ± 0%  -21.24%  (p=0.008 n=5+5)
SimilaritySearch/numWorkers=8-10     73.7ms ± 0%    57.9ms ± 0%  -21.37%  (p=0.008 n=5+5)
SimilaritySearch/numWorkers=16-10    83.6ms ± 2%    65.7ms ± 2%  -21.38%  (p=0.008 n=5+5)
```

## Test plan

Included benchmark results in PR. Added a quicktest to validate that, for all inputs, `CosineSimilarity` behaves the same as the naive implementation. 

<!-- All pull requests REQUIRE a test plan: https://docs.sourcegraph.com/dev/background-information/testing_principles -->